requirements.txt
tabulate
scipy==1.11.4
matplotlib==3.8.2
scikit-learn==1.3.2
timm==0.9.12
tqdm==4.66.1
onnxruntime-gpu==1.16.3
shapely==2.0.2
opencv-python==4.9.0.80
pandas==2.1.4
natsort==8.4.0
GPUtil==1.4.0
click>= 8.0.3
# cudatoolkit>= 10.2
faiss-cpu
#faiss-gpu
matplotlib>= 3.5.0
pillow>= 8.4.0
pretrainedmodels>= 0.7.4
torch>= 1.10.0
scikit-image>= 0.18.3
scikit-learn>= 1.0.1
scipy>= 1.7.1
torchvision>= 0.11.1
tqdm>= 4.62.3